feat: support web SSO login
After the Web SSO login is successfully configured on the back-end, you can select a login mode on the login page based on the configuration. The default login mode is Keystone authentication. If you select the Web SSO login mode, click OK to switch to the Web SSO configuration page.

Implements: blueprint skyline-sso-oid
Change-Id: I438adb31c758287525ba896b4d2b21d00842c3e7